Removed TestEntity from tests
Refactored in two classes, one for frontend, one for backend
