+++ schoolmanager-backend/src/test/java/apt/project/backend/domain/TestEntity.java
@@ -1,14 +1,16 @@
-package apt.project.backend.repository;
+package apt.project.backend.domain;
 
 import javax.persistence.Entity;
 
-import apt.project.backend.domain.BaseEntity;
-
 @Entity
-class TestEntity extends BaseEntity {
+public class TestEntity extends BaseEntity {
     // This entity has the only purpose to represent a concrete class
     // extending BaseEntity
+
     private String aField;
+
+    public TestEntity() {
+    }
 
     public String getaField() {
         return aField;
