+++ schoolmanager-backend/src/test/java/apt/project/backend/repository/BaseRepositoryTest.java
@@ -14,6 +14,8 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import apt.project.backend.domain.TestEntity;
+
 public class BaseRepositoryTest {
 
     private static EntityManagerFactory entityManagerFactory;
